_layouts from app to src folder, renamed layouts
