+++ src/app/sign-in/page.tsx
@@ -1,5 +1,5 @@
 import { SignInForm } from "../_components";
-import { PageSection } from "../_layouts/pageSection/PageSection";
+import { PageSection } from "../../layouts/pageSection/PageSection";
 
 export default function SignInPage() {
   return (
